new Version (number) and revisionDesc
The Version number in .conf file goes from 2.0.0 to 2.1.0.

The revisionDesc in header.xml added with date 2023.08.15.

The contents of header.xml is built into mapm.osis.xml.
I.e., we make the change in header.xml and that change
propagates to mapm.osis.xml when we build mapm.osis.xml.